Remove type casting for unit testing

+++ src/wp-includes/post-thumbnail-template.php
@@ -37,7 +37,7 @@ function get_post_thumbnail_id( $post = null ) {
 		return '';
 	}
 
-	$thumbnail_id = (int) get_post_meta( $post->ID, '_thumbnail_id', true );
+	$thumbnail_id = get_post_meta( $post->ID, '_thumbnail_id', true );
 
 	/**
 	 * Filters post thumbnail ID.
@@ -47,7 +47,7 @@ function get_post_thumbnail_id( $post = null ) {
 	 * @param int|false        $thumbnail_id Post thumbnail ID or false if the post does not exist.
 	 * @param int|WP_Post|null $post         Post ID or WP_Post object. Default is global `$post`.
 	 */
-	return (int) apply_filters( 'post_thumbnail_id', $thumbnail_id, $post );
+	return apply_filters( 'post_thumbnail_id', $thumbnail_id, $post );
 }
 
 /**
